requirements.txt
streamlit
requests
streamlit_lottie
python-dotenv